Add functions to strip whitespace
In ed, commands like "   a    \n" are equivalent to "a\n", so we need
a function that can merge them. Similarly, since every line ends with
a newline, it's redundant to store them, and performing string
comparisons that include them are slightly more annoying, so we need
a function to strip those.