Allow test suite to continue on failures
Stopping on the first failure is annoying behavior. This commit forks
the test runner to protect the suite from each test, so a SEGFAULT in
one won't crash all the tests.

Move all of the strings onto the stack so that they can be modified in
place without crashing.

Provide better reporting of which tests pass.

Change the Makefile to run the compiler with more strict flags
(e.g. -Wall -Werror)