Fix lstrip
The lstrip function was copying a fixed size string, instead of
checking the string length. It also might have been damaging the
null terminator on the string.

Fix test_rstrip

The rstrip test had a typo in the test string, so the test was failing
even though the function worked correctly.