Add support for reading files
The `r` command with no arguments will read from the current file being
used for writing and place all the lines after the current point in the
buffer. The `r` command with an argument will update the current name
the same way that `w <name>` does, and then will read.

NOTE: I don't think this quite matches the behavior of ed, which will
by default read to the end of the file. Since there is no support for
specifying addresses yet, this way is more flexible, but should be
changed when addresses are added.